Hash shard metadata
